BoxParser and FragmentedBoxParser do the same job, so use only BoxParser.
