+++ src/streaming/utils/BoxParser.js
@@ -29,6 +29,7 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
+import Debug from '../../core/Debug';
 import IsoFile from './IsoFile';
 import FactoryMaker from '../../core/FactoryMaker';
 import ISOBoxer from 'codem-isoboxer';
@@ -37,8 +38,13 @@ import IsoBoxSearchInfo from '../vo/IsoBoxSearchInfo';
 
 function BoxParser(/*config*/) {
 
-    let instance;
+    let logger,
+        instance;
     let context = this.context;
+
+    function setup() {
+        logger = Debug(context).getInstance().getLogger(instance);
+    }
 
     /**
      * @param {ArrayBuffer} data
@@ -108,6 +114,103 @@ function BoxParser(/*config*/) {
         return boxInfo;
     }
 
+    function getSamplesInfo(ab) {
+        if (!ab || ab.byteLength === 0) {
+            return {sampleList: [], lastSequenceNumber: NaN, totalDuration: NaN, numSequences: NaN};
+        }
+        let isoFile = parse(ab);
+        // zero or more moofs
+        let moofBoxes = isoFile.getBoxes('moof');
+        // exactly one mfhd per moof
+        let mfhdBoxes = isoFile.getBoxes('mfhd');
+
+        let sampleDuration,
+            sampleCompositionTimeOffset,
+            sampleCount,
+            sampleSize,
+            sampleDts,
+            sampleList,
+            sample,
+            i, j, k, l, m, n,
+            dataOffset,
+            lastSequenceNumber,
+            numSequences,
+            totalDuration;
+
+        numSequences = isoFile.getBoxes('moof').length;
+        lastSequenceNumber = mfhdBoxes[mfhdBoxes.length - 1].sequence_number;
+        sampleCount = 0;
+
+        sampleList = [];
+        let subsIndex = -1;
+        let nextSubsSample = -1;
+        for (l = 0; l < moofBoxes.length; l++) {
+            let moofBox = moofBoxes[l];
+            // zero or more trafs per moof
+            let trafBoxes = moofBox.getChildBoxes('traf');
+            for (j = 0; j < trafBoxes.length; j++) {
+                let trafBox = trafBoxes[j];
+                // exactly one tfhd per traf
+                let tfhdBox = trafBox.getChildBox('tfhd');
+                // zero or one tfdt per traf
+                let tfdtBox = trafBox.getChildBox('tfdt');
+                sampleDts = tfdtBox.baseMediaDecodeTime;
+                // zero or more truns per traf
+                let trunBoxes = trafBox.getChildBoxes('trun');
+                // zero or more subs per traf
+                let subsBoxes = trafBox.getChildBoxes('subs');
+                for (k = 0; k < trunBoxes.length; k++) {
+                    let trunBox = trunBoxes[k];
+                    sampleCount = trunBox.sample_count;
+                    dataOffset = (tfhdBox.base_data_offset || 0) + (trunBox.data_offset || 0);
+
+                    for (i = 0; i < sampleCount; i++) {
+                        sample = trunBox.samples[i];
+                        sampleDuration = (sample.sample_duration !== undefined) ? sample.sample_duration : tfhdBox.default_sample_duration;
+                        sampleSize = (sample.sample_size !== undefined) ? sample.sample_size : tfhdBox.default_sample_size;
+                        sampleCompositionTimeOffset = (sample.sample_composition_time_offset !== undefined) ? sample.sample_composition_time_offset : 0;
+                        let sampleData = {
+                            'dts': sampleDts,
+                            'cts': (sampleDts + sampleCompositionTimeOffset),
+                            'duration': sampleDuration,
+                            'offset': moofBox.offset + dataOffset,
+                            'size': sampleSize,
+                            'subSizes': [sampleSize]
+                        };
+                        if (subsBoxes) {
+                            for (m = 0; m < subsBoxes.length; m++) {
+                                let subsBox = subsBoxes[m];
+                                if (subsIndex < (subsBox.entry_count - 1) && i > nextSubsSample) {
+                                    subsIndex++;
+                                    nextSubsSample += subsBox.entries[subsIndex].sample_delta;
+                                }
+                                if (i == nextSubsSample) {
+                                    sampleData.subSizes = [];
+                                    let entry = subsBox.entries[subsIndex];
+                                    for (n = 0; n < entry.subsample_count; n++) {
+                                        sampleData.subSizes.push(entry.subsamples[n].subsample_size);
+                                    }
+                                }
+                            }
+                        }
+                        sampleList.push(sampleData);
+                        dataOffset += sampleSize;
+                        sampleDts += sampleDuration;
+                    }
+                }
+                totalDuration = sampleDts - tfdtBox.baseMediaDecodeTime;
+            }
+        }
+        return {sampleList: sampleList, lastSequenceNumber: lastSequenceNumber, totalDuration: totalDuration, numSequences: numSequences};
+    }
+
+    function getMediaTimescaleFromMoov(ab) {
+        let isoFile = parse(ab);
+        let mdhdBox = isoFile ? isoFile.getBox('mdhd') : undefined;
+
+        return mdhdBox ? mdhdBox.timescale : NaN;
+    }
+
     function parseUint32(data, offset) {
         return data[offset + 3] >>> 0 |
             (data[offset + 2] << 8) >>> 0 |
@@ -124,8 +227,12 @@ function BoxParser(/*config*/) {
 
     instance = {
         parse: parse,
-        findLastTopIsoBoxCompleted: findLastTopIsoBoxCompleted
+        findLastTopIsoBoxCompleted: findLastTopIsoBoxCompleted,
+        getMediaTimescaleFromMoov: getMediaTimescaleFromMoov,
+        getSamplesInfo: getSamplesInfo,
     };
+
+    setup();
 
     return instance;
 }
